+++ src/components/react/do/do-item-react.jsx
@@ -5,7 +5,7 @@ import RadiobuttonReact from '../basic/radiobutton-react.jsx';
 import NumericReact from '../basic/numeric-react.jsx';
 // import POTextileAutoSuggestReact from '../auto-suggests/po-textile-auto-suggest-react.jsx';
 import ProductAutoSuggestReact from '../auto-suggests/product-auto-suggest-react.jsx';
-import PoExternalAutoSuggestReact from '../auto-suggests/po-external-auto-suggest-react.jsx';
+import PoExternalAutoSuggestReact from '../auto-suggests/po-external-posted-auto-suggest-react.jsx';
 import UomAutoSuggestReact from '../auto-suggests/uom-auto-suggest-react.jsx';
 import DoItemFulfillmentReact from './do-item-fulfillment-react.jsx';
