--- src/modules/purchasing/po-textile/create-podl.html
@@ -1,12 +1,0 @@
-<template>
-    <h1 class="page-header">Tambah Eksternal PO Tekstil</h1>
-    <require from="./data-form-podl"></require>
-    <block>
-        <data-form-podl data.bind="data" error.bind="error"></data-form-podl>
-        <hr/>
-        <div class="btn-group pull-right">
-            <button class="btn btn-default" click.delegate="back()">Kembali</button>
-            <button class="btn btn-success" click.delegate="save()">Simpan</button>
-        </div>
-    </block>
-</template>
