--- src/modules/purchasing/po-textile/create.html
@@ -1,12 +1,0 @@
-<template>
-    <h1 class="page-header">Tambah Internal PO Tekstil</h1>
-    <require from="./data-form"></require>
-    <block>
-        <data-form data.bind="data" error.bind="error"></data-form>
-        <hr/>
-        <div class="btn-group pull-right">
-            <button class="btn btn-default" click.delegate="back()">Kembali</button>
-            <button class="btn btn-success" click.delegate="save()">Simpan</button>
-        </div>
-    </block>
-</template>
