--- src/modules/purchasing/po-textile/edit-podl.html
@@ -1,12 +1,0 @@
-<template>
-    <h1 class="page-header">Ubah Eksternal PO Tekstil</h1>
-    <require from="./data-form-podl"></require>
-    <block>
-        <data-form-podl data.bind="data" error.bind="error"></data-form-podl>
-        <hr/>
-        <div class="pull-right">
-            <button class="btn btn-default" click.delegate="view()">Kembali</button>
-            <button class="btn btn-success" click.delegate="save()">Simpan</button>
-        </div>
-    </block>
-</template>
