--- src/modules/purchasing/po-textile/edit-podl.js
@@ -1,30 +1,0 @@
-import {inject, Lazy} from 'aurelia-framework';
-import {Router} from 'aurelia-router';
-import {Service} from './service';
-
-
-@inject(Router, Service)
-export class Edit {
-    constructor(router, service) {
-        this.router = router;
-        this.service = service;
-    }
-
-    async activate(params) {
-        var id = params.id;
-        this.data = await this.service.getById(id);
-        console.log(this.data);
-    }
-
-    view() { 
-        this.router.navigateToRoute('view', { id: this.data._id });
-    }
-
-    save() {
-        this.service.update(this.data).then(result => {
-            this.view();
-        }).catch(e => {
-            this.error = e;
-        })
-    }
-}
