--- src/modules/purchasing/po-textile/edit.html
@@ -1,14 +1,0 @@
-<template>
-    <h3 if.bind="!data.isSplit">Ubah Internal PO Tekstil</h1>
-    <h3 if.bind="data.isSplit">Pecah Internal PO Tekstil</h1>
-    <require from="./data-form"></require>
-    <block>
-        <data-form data.bind="data" error.bind="error"></data-form>
-        <hr/>
-        <div class="pull-right">
-            <button class="btn btn-default" click.delegate="view()">Kembali</button>
-            <button if.bind="!data.isSplit" class="btn btn-success" click.delegate="save()">Simpan</button>
-            <button if.bind="data.isSplit" class="btn btn-success" click.delegate="split()">Pecah PO</button>
-        </div>
-    </block>
-</template>
