--- src/modules/purchasing/po-textile/index.html
@@ -1,5 +1,0 @@
-<template> 
-  <div> 
-    <router-view></router-view>
-  </div>
-</template>
